requirements/base.txt
# Django
Django==1.10.7

# Environmental var configuration
django-configurations==2.0

# For the persistence stores
psycopg2==2.7.1
dj-database-url==0.4.2

# Models
django-model-utils==3.0.0

# Rest apis
djangorestframework==3.6.2
Markdown==2.6.8
django-filter==1.0.2

# Static files
whitenoise==3.3.0

# Media files
boto==2.46.1
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.7.0

# For asynchronous queuing
django-rq==0.9.5

# Time zones support
pytz==2017.2

django_nose
django-rest-swagger==2.1.2
djangorestframework-jwt
